requirements.txt
djangorestframework
django-filter
drf-flex-fields
djangorestframework-simplejwt
redis